backend/requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
transformers
torch
fastapi
uvicorn
pydantic
streamlit
numpy
pandas
tqdm
ragas
datasets
python-dotenv
pypdf
huggingface_hub
openai